+++ engine/samples/physics/main.cpp
@@ -1,6 +1,5 @@
 #include <cubos/engine/collisions/plugin.hpp>
 #include <cubos/engine/fixed_step/plugin.hpp>
-#include <cubos/engine/fixed_substep/plugin.hpp>
 #include <cubos/engine/physics/plugin.hpp>
 #include <cubos/engine/physics/plugins/gravity.hpp>
 #include <cubos/engine/physics/solver/plugin.hpp>
@@ -40,7 +39,6 @@ int main(int argc, char** argv)
     cubos.plugin(voxelsPlugin);
     cubos.plugin(fixedStepPlugin);
     cubos.plugin(collisionsPlugin);
-    cubos.plugin(fixedSubstepPlugin);
     cubos.plugin(physicsPlugin);
     cubos.plugin(solverPlugin);
     cubos.plugin(gravityPlugin);
