+++ engine/src/physics/solver/plugin.cpp
@@ -1,11 +1,9 @@
 #include "penetration_constraint/plugin.hpp"
 
-#include <glm/glm.hpp>
-
-#include <cubos/engine/fixed_step/plugin.hpp>
-#include <cubos/engine/fixed_substep/plugin.hpp>
 #include <cubos/engine/physics/plugin.hpp>
 #include <cubos/engine/physics/solver/plugin.hpp>
+
+#include "../fixed_substep/plugin.hpp"
 
 CUBOS_DEFINE_TAG(cubos::engine::physicsSolveTag);
 
@@ -13,8 +11,6 @@ using namespace cubos::engine;
 
 void cubos::engine::solverPlugin(Cubos& cubos)
 {
-    cubos.depends(fixedStepPlugin);
-    cubos.depends(fixedSubstepPlugin);
     cubos.depends(physicsPlugin);
 
     cubos.tag(physicsSolveTag)
